Move regenerate client secret from general settings tab.
